Fix custom video player issues
Replaced loading screen while player is not ready with a better and working animation. Volume
bar no longer blocks the seek bar and css has been improved.
Added a settings menu to change to various stream quality.

To be done: Stream quality should include default and the standard qualities only. Also need to improve
the CSS of the settings menu.